Fixing PC Deposit/Withdraw Glitch
When depositing/withdrawing items in the PC, the game would wrongly assume the item selected was the last one on the visible list. This would cause all sort of glitchy behavior, like being able to deposit more items than you had on hand, causing duplication glitches or loss of items.